requirements.txt
dotenv~=0.9.9
python-dotenv~=1.0.1
langchain-google-genai~=2.0.10
pydantic~=2.10.6
protobuf~=5.29.4
langgraph~=0.3.20
langgraph-supervisor~=0.0.11
requests~=2.32.3
uvicorn~=0.30.6
fastapi~=0.115.3
bs4~=0.0.2
beautifulsoup4~=4.13.3
gradio~=5.25.1
vertexai~=1.71.1
google-cloud-aiplatform~=1.71.1
google-generativeai
google-genai
agno~=1.5.5
groq~=0.21.0
chromadb~=1.0.4
pandas~=2.2.3
fonttools~=4.56.0
pypandoc~=1.15
smolagents~=1.16.1
langchain-community~=0.3.20
pygeai~=0.2.6
openai~=1.68.2
supabase~=2.15.3
playwright~=1.53.0
pyppeteer~=1.0.2
Flask~=2.3.3
Flask-Cors~=4.0.0
psycopg_binary
arize-phoenix